refactor: move save_fitness, save_best_genome and load_best_genome to utils.py
